+++ Sources/Spine/Extensions/Float+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  Float+Extensions.swift
 //  
 //
 //  Created by Max Gribov on 19.11.2022.
